+++ app/client/views/search.js
@@ -2,8 +2,8 @@ const html = require('choo/html')
 const css = require('csjs-inject')
 const C = require('../lib/constants')
 const imgpath = require('../lib/imgpath')
-const throttle = require('lodash/throttle')
 const equal = require('lodash/isEqual')
+const CacheComponent = require('cache-component')
 
 const debug = require('debug')('sciencefair:view:search')
 
@@ -29,20 +29,6 @@ const style = css`
   display: flex;
 }
 
-.input {
-  width: 100%;
-  height: 30px;
-  border: none;
-  border-bottom: dotted 2px ${C.DARKBLUE};
-  font-size: 130%;
-  padding-left: 40px;
-  padding-bottom: 5px;
-  font-family: CooperHewitt-Book;
-  background: none;
-  display: flex;
-  outline: none;
-}
-
 .img {
   position: absolute;
   left: 2px;
@@ -51,68 +37,26 @@ const style = css`
   z-index: 900;
 }
 
-.tags {
-  position: absolute;
-  right: 34px;
-}
-
-.clear {
-  position: absolute;
-  padding: 5px;
-  right: 8px;
-  top: 0;
-  width: 20px;
-  height: 30px;
-  background-color: ${C.DARKBLUE};
-  -webkit-mask: url(${imgpath('close.svg')}) center / contain no-repeat;
-}
-
 `
-
-const CacheComponent = require('cache-component')
 
 function CachedSearch () {
   if (!(this instanceof CachedSearch)) return new CachedSearch()
-  this._color = null
   CacheComponent.call(this)
 }
 CachedSearch.prototype = Object.create(CacheComponent.prototype)
 
 CachedSearch.prototype._render = function (state, emit) {
-  this._searchstate
-
-  const emitify = throttle(emit, 200, { leading: true })
-  const placeholder = getplaceholder(state)
-
-  const clearing = state.search.clearing
-  const querystring = state.search.querystring
-  const query = state.search.query
-  const tagquery = state.search.tagquery
-  const tags = state.search.tags
-
-  const hasquery = !!query
-  const hastagquery = !!tagquery
-  const hastags = tags.length > 0
-
-  const input =
-
-  const tagbadges = tags.length ? html`
-
-  <div class="${style.tags}">
-    ${tags.map(tag => require('./search_tag')(tag, emit))}
-  </div>
-
-  ` : null
+  this._searchstate = state
 
   return html`
 
   <div class="${style.search}">
     <div class="${style.wrapper}">
-      ${require('./search_input')(state, emit)}
-      ${tagbadges}
-      ${require('./search_clearbtn')(state, emit)}
       <img class="${style.img}" src="${imgpath('search.svg')}" />
-      ${autocomplete()}
+      ${require('./search_input')(state, emit)}
+      ${require('./search_tags')(state, emit)}
+      ${require('./search_clearbtn')(state, emit)}
+      ${require('./autocomplete')(state, emit)}
     </div>
   </div>
